Move heal types to separate files
This patch moves the code that implements the different heal types to
separate files.
