+++ cmd/alert-healer/alerts_worker.go
@@ -19,20 +19,14 @@ package main
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"regexp"
 	"text/template"
 
 	"github.com/golang/glog"
-	batch "k8s.io/api/batch/v1"
-	core "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/runtime"
 
 	alertmanager "github.com/jhernand/openshift-monitoring/pkg/alertmanager"
 	monitoring "github.com/jhernand/openshift-monitoring/pkg/apis/monitoring/v1alpha1"
-	awx "github.com/jhernand/openshift-monitoring/pkg/awx"
 )
 
 func (h *Healer) runAlertsWorker() {
@@ -241,306 +235,6 @@ func (h *Healer) runAction(rule *monitoring.HealingRule, action *monitoring.Heal
 			"There are no action details, rule '%s' will have no effect on alert '%s'",
 			rule.ObjectMeta.Name,
 			alert.Labels["alertname"],
-		)
-	}
-
-	return nil
-}
-
-func (h *Healer) runAWXJob(rule *monitoring.HealingRule, action *monitoring.AWXJobAction, alert *alertmanager.Alert) error {
-	glog.Infof(
-		"Running AWX job from project '%s' and template '%s' to heal alert '%s'",
-		action.Project,
-		action.Template,
-		alert.Labels["alertname"],
-	)
-
-	// Load the AWX credentials:
-	secret := action.SecretRef
-	if secret == nil {
-		return fmt.Errorf("The secret containing the AWX credentials hasn't been specified")
-	}
-	username, password, err := h.loadAWXSecret(rule, secret)
-	if err != nil {
-		return err
-	}
-
-	// Create the connection to the AWX server:
-	connection, err := awx.NewConnectionBuilder().
-		Url(action.Address).
-		Proxy(action.Proxy).
-		Username(username).
-		Password(password).
-		Insecure(true).
-		Build()
-	if err != nil {
-		return err
-	}
-	defer connection.Close()
-
-	// Retrieve the job template:
-	templatesResource := connection.JobTemplates()
-	templatesResponse, err := templatesResource.Get().
-		Filter("project__name", action.Project).
-		Filter("name", action.Template).
-		Send()
-	if err != nil {
-		return err
-	}
-	if templatesResponse.Count() == 0 {
-		return err
-	}
-
-	// Launch the jobs:
-	for _, template := range templatesResponse.Results() {
-		err := h.launchAWXJob(connection, template, action)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (h *Healer) launchAWXJob(connection *awx.Connection, template *awx.JobTemplate, action *monitoring.AWXJobAction) error {
-	templateId := template.Id()
-	templateName := template.Name()
-	launchResource := connection.JobTemplates().Id(templateId).Launch()
-	_, err := launchResource.Post().
-		ExtraVars(action.ExtraVars).
-		Send()
-	if err != nil {
-		return err
-	}
-	glog.Infof(
-		"Request to launch AWX job from template '%s' has been sent",
-		templateName,
-	)
-	return nil
-}
-
-func (h *Healer) loadAWXSecret(rule *monitoring.HealingRule, reference *core.SecretReference) (username, password string, err error) {
-	var data []byte
-	var ok bool
-
-	// The name of the secret is mandatory:
-	name := reference.Name
-	if name == "" {
-		err = fmt.Errorf(
-			"Can't load AWX secret for rule '%s', the name hasn't been specified",
-			rule.ObjectMeta.Name,
-		)
-		return
-	}
-
-	// The namespace of the secret is optional, the default is the namespace of the rule:
-	namespace := reference.Namespace
-	if namespace == "" {
-		namespace = rule.ObjectMeta.Namespace
-	}
-
-	// Retrieve the secret:
-	resource := h.k8sClient.CoreV1().Secrets(namespace)
-	secret, err := resource.Get(name, meta.GetOptions{})
-	if err != nil {
-		err = fmt.Errorf(
-			"Can't load secret '%s' from namespace '%s': %s",
-			name,
-			namespace,
-			err.Error(),
-		)
-		return
-	}
-
-	// Extract the user name:
-	data, ok = secret.Data["username"]
-	if !ok {
-		err = fmt.Errorf(
-			"Secret '%s' from namespace '%s' doesn't contain the 'username' entry",
-			name,
-			namespace,
-		)
-		return
-	}
-	username = string(data)
-
-	// Extract the password:
-	data, ok = secret.Data["password"]
-	if !ok {
-		err = fmt.Errorf(
-			"Secret '%s' from namespace '%s' doesn't contain the 'password' entry",
-			name,
-			namespace,
-		)
-		return
-	}
-	password = string(data)
-
-	return
-}
-
-func (h *Healer) runBatchJob(rule *monitoring.HealingRule, job *batch.Job, alert *alertmanager.Alert) error {
-	glog.Infof(
-		"Running batch job '%s' to heal alert '%s'",
-		job.ObjectMeta.Name,
-		alert.Labels["alertname"],
-	)
-
-	// The name of the job is mandatory:
-	name := job.ObjectMeta.Name
-	if name == "" {
-		return fmt.Errorf(
-			"Can't create job for rule '%s', the name hasn't been specified",
-			rule.ObjectMeta.Name,
-		)
-	}
-
-	// The namespace of the job is optional, the default is the namespace of the rule:
-	namespace := job.ObjectMeta.Namespace
-	if namespace == "" {
-		namespace = rule.ObjectMeta.Namespace
-	}
-
-	// Get the resource that manages the collection of batch jobs:
-	resource := h.k8sClient.Batch().Jobs(namespace)
-
-	// Try to create the job:
-	job = job.DeepCopy()
-	job.ObjectMeta.Name = name
-	job.ObjectMeta.Namespace = namespace
-	_, err := resource.Create(job)
-	if errors.IsAlreadyExists(err) {
-		glog.Warningf(
-			"Batch job '%s' already exists, will do nothing to heal alert '%s'",
-			job.ObjectMeta.Name,
-			alert.Labels["alertname"],
-		)
-	} else if err != nil {
-		return err
-	} else {
-		glog.Infof(
-			"Batch job '%s' to heal alert '%s' has been created",
-			job.ObjectMeta.Name,
-			alert.Labels["alertname"],
-		)
-	}
-
-	return nil
-}
-
-func (h *Healer) runAnsiblePlaybook(rule *monitoring.HealingRule, action *monitoring.AnsiblePlaybookAction, alert *alertmanager.Alert) error {
-	glog.Infof(
-		"Running Ansible playbook from healing rule '%s' and alert '%s'",
-		rule.ObjectMeta.Name,
-		alert.Labels["alertname"],
-	)
-
-	// The configuration map and the job will be in the same namespace and will have the same name
-	// than the alert:
-	namespace := alert.Labels["namespace"]
-	name := alert.Labels["alertname"]
-
-	// Populate the configuration map:
-	config := &core.ConfigMap{
-		ObjectMeta: meta.ObjectMeta{
-			Namespace: namespace,
-			Name:      name,
-		},
-		Data: map[string]string{
-			"playbook.yml": action.Playbook,
-			"inventory":    action.Inventory,
-		},
-	}
-
-	// Create the configuration map:
-	configsResource := h.k8sClient.Core().ConfigMaps(namespace)
-	_, err := configsResource.Create(config)
-	if errors.IsAlreadyExists(err) {
-		glog.Infof(
-			"Configuration map '%s' already exists",
-			config.ObjectMeta.Name,
-		)
-	} else if err != nil {
-		return err
-	} else {
-		glog.Infof(
-			"Created configuration map '%s'",
-			config.ObjectMeta.Name,
-		)
-	}
-
-	// Determine the user that will be used to run Ansible. This is needed because the UID must
-	// match the permissions set inside the image, otherwise Ansible will not be able to write to
-	// the home directory.
-	uid := int64(1000000000)
-
-	// Build the Ansible command line:
-	command := []string{
-		"/usr/bin/ansible-playbook",
-		"--inventory=/config/inventory",
-	}
-	if action.ExtraVars != "" {
-		command = append(command, "--extra-vars="+action.ExtraVars)
-	}
-	command = append(command, "/config/playbook.yml")
-
-	// Populate the job:
-	job := &batch.Job{
-		ObjectMeta: meta.ObjectMeta{
-			Namespace: namespace,
-			Name:      name,
-		},
-		Spec: batch.JobSpec{
-			Template: core.PodTemplateSpec{
-				Spec: core.PodSpec{
-					Volumes: []core.Volume{
-						core.Volume{
-							Name: "config",
-							VolumeSource: core.VolumeSource{
-								ConfigMap: &core.ConfigMapVolumeSource{
-									LocalObjectReference: core.LocalObjectReference{
-										Name: config.ObjectMeta.Name,
-									},
-								},
-							},
-						},
-					},
-					Containers: []core.Container{
-						core.Container{
-							Name:  "ansible-runner",
-							Image: "openshift-monitoring/ansible-runner:0.0.0",
-							SecurityContext: &core.SecurityContext{
-								RunAsUser: &uid,
-							},
-							Command: command,
-							VolumeMounts: []core.VolumeMount{
-								core.VolumeMount{
-									Name:      "config",
-									MountPath: "/config",
-								},
-							},
-						},
-					},
-					RestartPolicy: core.RestartPolicyNever,
-				},
-			},
-		},
-	}
-
-	// Create the job:
-	jobsResource := h.k8sClient.Batch().Jobs(namespace)
-	_, err = jobsResource.Create(job)
-	if errors.IsAlreadyExists(err) {
-		glog.Infof(
-			"Job '%s' already exists",
-			job.ObjectMeta.Name,
-		)
-	} else if err != nil {
-		return err
-	} else {
-		glog.Infof(
-			"Created job '%s'",
-			job.ObjectMeta.Name,
 		)
 	}
